Refactored ui and database code into pacakges and updated build.gradle file.
