+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:theme="@style/Theme.ProfileManager"
         tools:targetApi="31">
         <activity
-            android:name=".CreateProfileScreen"
+            android:name=".ui.createprofile.CreateProfileScreen"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -21,7 +21,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".AllProfiles"
+            android:name=".ui.allprofiles.AllProfilesScreen"
             android:exported="true" />
     </application>
