+++ app/src/main/java/com/example/profilemanager/database/UserDao.kt
@@ -1,6 +1,7 @@
-package com.example.profilemanager
+package com.example.profilemanager.database
 
 import androidx.room.*
+import com.example.profilemanager.database.User
 
 @Dao
 interface UserDao {
