+++ app/src/main/java/com/example/profilemanager/ui/allprofiles/AllProfilesContract.kt
@@ -1,13 +1,12 @@
-package com.example.profilemanager
+package com.example.profilemanager.ui.allprofiles
+
+import com.example.profilemanager.database.User
 
 interface AllProfilesContract {
     interface View {
-
     }
 
     interface Presenter {
         fun getAllUsers(): List<User>
-
-        fun onDestroy()
     }
 }
