+++ app/src/main/res/values/strings.xml
@@ -8,7 +8,7 @@
     <string name="female">F</string>
     <string name="age">Age</string>
     <string name="date_of_birth_hint">Date of Birth (mm/dd/yyyy)</string>
-    <string name="brief_desc_text">Add a brief descirption about yourself</string>
+    <string name="brief_desc_text">Add a brief description about yourself</string>
     <string name="submit">Submit</string>
     <string name="view_all_profiles">View All Profiles</string>
     <string name="personal_info">First: %s Last: %s Age: %s DOB: %s Gender: %s Education: %s</string>
